Add Unit Testing for Error handling, add dependencies for testing and reorganize project hierarchy
