--- src/main/java/com/adobe/bookstore/GlobalExceptionHandler.java
@@ -1,16 +1,0 @@
-package com.adobe.bookstore;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.apache.logging.log4j.Logger;
-
-@ControllerAdvice
-
-public class GlobalExceptionHandler {
-    @ExceptionHandler(OrderNotFoundException.class)
-    public ResponseEntity<String> handleOrderNotFoundException(OrderNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-}
