--- src/main/java/com/adobe/bookstore/OrderNotFoundException.java
@@ -1,7 +1,0 @@
-package com.adobe.bookstore;
-
-public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(String id) {
-        super("Order " + id + " not found");
-    }
-}
