+++ src/main/java/com/adobe/error/OrderNotFoundException.java
@@ -1,0 +1,8 @@
+package com.adobe.error;
+
+public class OrderNotFoundException extends RuntimeException {
+
+  public OrderNotFoundException(String id) {
+    super("Order " + id + " not found");
+  }
+}
